Processing Files for Postgres Output
This includes the processing files and outputs for both Python and R files to be used as inputs for the ConvAE architecture. Due to random states, the training and test sets will be different. Additionally, since the processing is done on a fabricated dataset that do not have the distributions of true prescription data, the results will not be as strong as the original MIMIC data's results. Regardless, the pipleine can be run using this.

The R or python processing scripts can be used to prepare the input data.